backend-v2/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
jsonschema==4.25.1
httpx==0.28.1
pytest==8.4.2
